requirements.txt
protobuf>=3.19.0
pycryptodome~=3.20.0
google>=3.0.0
tencentcloud-sdk-python>=3.0.1224
